Fixed v3 people user deletion command

* Fixed v3 people user deletion command

* People deletion v3 fix


Approved-by: Aruna Velayutham
Approved-by: Vandhana sridhar

+++ src/main/java/com/facilio/bmsconsoleV3/commands/DeletePeopleUserPortalAccessV3.java
@@ -1,37 +1,22 @@
 package com.facilio.bmsconsoleV3.commands;
 
-import com.facilio.bmsconsoleV3.context.V3PeopleContext;
-import com.facilio.bmsconsoleV3.context.V3TenantContactContext;
-import com.facilio.bmsconsoleV3.context.V3VendorContactContext;
 import com.facilio.bmsconsoleV3.util.V3PeopleAPI;
 import com.facilio.command.FacilioCommand;
 import com.facilio.v3.context.Constants;
 import org.apache.commons.chain.Context;
+import org.apache.commons.collections.CollectionUtils;
 
 import java.util.List;
-import java.util.Map;
 
 public class DeletePeopleUserPortalAccessV3 extends FacilioCommand {
     @Override
     public boolean executeCommand(Context context) throws Exception {
-        String moduleName = Constants.getModuleName(context);
-        Map<String, List> recordMap = (Map<String, List>) context.get(Constants.RECORD_MAP);
-        List<V3PeopleContext> peopleList = recordMap.get(moduleName);
-        for (V3PeopleContext people : peopleList) {
-            List records = recordMap.get(moduleName);
-            long id =0;
-            if (records != null && !records.isEmpty()) {
-                for (int i = 0; i < records.size(); i++) {
-                    if (people.getPeopleType() == V3PeopleContext.PeopleType.TENANT_CONTACT.getIndex()) {
-                        V3TenantContactContext tc = (V3TenantContactContext) records.get(i);
-                        id = tc.getId();
-                    } else if (people.getPeopleType() == V3PeopleContext.PeopleType.VENDOR_CONTACT.getIndex()) {
-                        V3VendorContactContext vc = (V3VendorContactContext) records.get(i);
-                        id = vc.getId();
-                    }
-                }
+
+        List<Long> peopleIds = Constants.getRecordIds(context);
+        if(CollectionUtils.isNotEmpty(peopleIds)) {
+            for (Long id : peopleIds) {
+                V3PeopleAPI.deletePeopleUsers(id);
             }
-            V3PeopleAPI.deletePeopleUsers(id);
         }
         return false;
     }
